Move Git functions to git-utils.ts
