+++ src/analyze-action-env.test.ts
@@ -5,6 +5,7 @@ import * as actionsUtil from "./actions-util";
 import * as analyze from "./analyze";
 import * as api from "./api-client";
 import * as configUtils from "./config-utils";
+import * as gitUtils from "./git-utils";
 import * as statusReport from "./status-report";
 import {
   setupTests,
@@ -31,7 +32,7 @@ test("analyze action with RAM & threads from environment variables", async (t) =
       .stub(statusReport, "createStatusReportBase")
       .resolves({} as statusReport.StatusReportBase);
     sinon.stub(statusReport, "sendStatusReport").resolves();
-    sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+    sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
 
     const gitHubVersion: util.GitHubVersion = {
       type: util.GitHubVariant.DOTCOM,
