+++ src/trap-caching.test.ts
@@ -14,6 +14,7 @@ import {
 } from "./codeql";
 import * as configUtils from "./config-utils";
 import { Feature } from "./feature-flags";
+import * as gitUtils from "./git-utils";
 import { Language } from "./languages";
 import { getRunnerLogger } from "./logging";
 import {
@@ -96,7 +97,7 @@ function getTestConfigWithTempDir(tempDir: string): configUtils.Config {
 test("check flags for JS, analyzing default branch", async (t) => {
   await util.withTmpDir(async (tmpDir) => {
     const config = getTestConfigWithTempDir(tmpDir);
-    sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+    sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
     const result = await getTrapCachingExtractorConfigArgsForLang(
       config,
       Language.javascript,
@@ -112,7 +113,7 @@ test("check flags for JS, analyzing default branch", async (t) => {
 test("check flags for all, not analyzing default branch", async (t) => {
   await util.withTmpDir(async (tmpDir) => {
     const config = getTestConfigWithTempDir(tmpDir);
-    sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(false);
+    sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(false);
     const result = await getTrapCachingExtractorConfigArgs(config);
     t.deepEqual(result, [
       `-O=javascript.trap.cache.dir=${path.resolve(tmpDir, "jsCache")}`,
@@ -139,7 +140,7 @@ test("get languages that support TRAP caching", async (t) => {
 test("upload cache key contains right fields", async (t) => {
   const loggedMessages = [];
   const logger = getRecordingLogger(loggedMessages);
-  sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+  sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
   sinon.stub(util, "tryGetFolderBytes").resolves(999_999_999);
   const stubSave = sinon.stub(cache, "saveCache");
   process.env.GITHUB_SHA = "somesha";
@@ -160,7 +161,7 @@ test("download cache looks for the right key and creates dir", async (t) => {
     const loggedMessages = [];
     const logger = getRecordingLogger(loggedMessages);
     sinon.stub(actionsUtil, "getTemporaryDirectory").returns(tmpDir);
-    sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(false);
+    sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(false);
     const stubRestore = sinon.stub(cache, "restoreCache").resolves("found");
     const eventFile = path.resolve(tmpDir, "event.json");
     process.env.GITHUB_EVENT_NAME = "pull_request";
@@ -200,8 +201,8 @@ test("cleanup removes only old CodeQL TRAP caches", async (t) => {
     // This config specifies that we are analyzing JavaScript and Ruby, but not Swift.
     const config = getTestConfigWithTempDir(tmpDir);
 
-    sinon.stub(actionsUtil, "getRef").resolves("refs/heads/main");
-    sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
+    sinon.stub(gitUtils, "getRef").resolves("refs/heads/main");
+    sinon.stub(gitUtils, "isAnalyzingDefaultBranch").resolves(true);
     const listStub = sinon.stub(apiClient, "listActionsCaches").resolves([
       // Should be kept, since it's not relevant to CodeQL. In reality, the API shouldn't return
       // this in the first place, but this is a defensive check.
